element menu was updated

+++ yew_app/src/lib.rs
@@ -38,8 +38,8 @@ use crate::fem::{FEModel, FEData};
 use crate::auxiliary::FEDrawnElementData;
 
 
-pub type ElementsNumbers = u32;
-pub type ElementsValues = f64;
+pub type ElementsNumbers = u16;
+pub type ElementsValues = f32;
 
 pub const TOLERANCE: ElementsValues = 1e-6;
